+++ src/test/java/io/github/mat3e/HelloServiceTest.java
@@ -56,7 +56,7 @@ public class HelloServiceTest {
     private LangRepository fallbackLangIdRepository() {
         return new LangRepository() {
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 if (id.equals(HelloService.FALLBACK_LANG.getId())) {
                     return Optional.of(new Lang(null, FALLBACK_ID_WELCOME, null));
                 }
@@ -68,7 +68,7 @@ public class HelloServiceTest {
         private LangRepository alwaysReturningHelloRepository () {
             return new LangRepository() {
                 @Override
-                Optional<Lang> findById(Long id) {
+                Optional<Lang> findById(Integer id) {
                     return Optional.of(new Lang(null, WELCOME, ""));
                 }
             };
